Moves `testAccPreCheck` to separate file
